Sync remotely on file attribute changes
Gio.FileMonitorEvent.CHANGED is only emitted when the file content
changes, but we also want to sync on modification time changes since the
rest of the syncing logic uses that. Sync on
Gio.FileMonitorEvent.ATTRIBUTE_CHANGED for that. It might trigger extra
syncs when other attributes (e.g., permissions) change, but better to
have too many syncs than not enough.